desloppify: relocate toast converters + extract close-worktree closure
- Move ToastConverters (Severity{Brush,Icon}, ProgressToWidth) from
  CodeScope.App/Toasts/ into CodeScope.Ui/Converters/. They depend
  only on ToastSeverity (already in Ui/Services/) so they belong
  alongside the other IValueConverter siblings. Update XAML xmlns
  refs in ToastView.xaml and DesignTokens.xaml to assembly-qualified
  clr-namespace; collapse the 4th SafeToken/Sanitize duplicate (in
  ToastActionViewModel.AutomationId) onto the shared
  Ui.AutomationIds.SafeToken helper. AutomationIds is promoted from
  internal to public so cross-assembly callers can reach it.
- Extract MainViewModel's 80-line inline CloseWorktreeSessionsAsync
  lambda (snapshot/close/rollback choreography) into private methods
  CloseWorktreeSessionsAsync + RestoreClosedWorktreeSessionsAsync.
  AttachSidebar now just hands the method group to the sidebar.
