+++ src/CodeScope.App/Styles/DesignTokens.xaml
@@ -17,7 +17,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:sys="clr-namespace:System;assembly=mscorlib"
-    xmlns:toasts="clr-namespace:NoScope.CodeScope.App.Toasts">
+    xmlns:toasts="clr-namespace:NoScope.CodeScope.Ui.Converters;assembly=CodeScope.Ui">
 
     <!-- ========================================================================= -->
     <!-- §2 COLOR PALETTE                                                           -->
